+++ Assets/Prefabs/Blocker Door .prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 131103acfa81f664c9fafc4dee7ca4a5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
