+++ Assets/Scenes/Test Scenes/Genric Test Scenes/Genric Door Test.unity
@@ -7712,7 +7712,7 @@ PrefabInstance:
     - target: {fileID: 6377068023124523544, guid: 0dc4edfe971154522b6efa571cdb6b0b,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -20.84214
+      value: -20.842133
       objectReference: {fileID: 0}
     - target: {fileID: 6377068023124523544, guid: 0dc4edfe971154522b6efa571cdb6b0b,
         type: 3}
@@ -12227,7 +12227,7 @@ PrefabInstance:
     - target: {fileID: 1951375231002363255, guid: 196469da6d87b9043b7a7261561efb1e,
         type: 3}
       propertyPath: m_TagString
-      value: Enemy
+      value: Wall
       objectReference: {fileID: 0}
     - target: {fileID: 7319345577122859348, guid: 196469da6d87b9043b7a7261561efb1e,
         type: 3}
@@ -16821,7 +16821,7 @@ GameObject:
   - component: {fileID: 1184002744}
   - component: {fileID: 1184002746}
   m_Layer: 0
-  m_Name: PlayerTriggerCheck
+  m_Name: PlayerTriggerArea
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -16979,6 +16979,127 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ac0b09e7857660247b1477e93731de29, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1 &1227815517
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1227815520}
+  - component: {fileID: 1227815519}
+  - component: {fileID: 1227815518}
+  - component: {fileID: 1227815521}
+  m_Layer: 0
+  m_Name: 'Blocker Door '
+  m_TagString: Wall
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1227815518
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1227815517}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5b3dbc667b5177a45b685a86862fa960, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!212 &1227815519
+SpriteRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1227815517}
+  m_Enabled: 1
+  m_CastShadows: 0
+  m_ReceiveShadows: 0
+  m_DynamicOccludee: 1
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 0
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_Sprite: {fileID: 21300000, guid: 858ea317411510c4790e5af8a4cdd907, type: 3}
+  m_Color: {r: 0.7607844, g: 0.7607844, b: 0.7607844, a: 1}
+  m_FlipX: 0
+  m_FlipY: 0
+  m_DrawMode: 0
+  m_Size: {x: 1, y: 1}
+  m_AdaptiveModeThreshold: 0.5
+  m_SpriteTileMode: 0
+  m_WasSpriteAssigned: 1
+  m_MaskInteraction: 0
+  m_SpriteSortPoint: 0
+--- !u!4 &1227815520
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1227815517}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0.21, y: 24.15, z: 0}
+  m_LocalScale: {x: 4.98, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 11
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!61 &1227815521
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1227815517}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 1, y: 1}
+    newSize: {x: 1, y: 1}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 1, y: 1}
+  m_EdgeRadius: 0
 --- !u!1 &1278591225
 GameObject:
   m_ObjectHideFlags: 0
@@ -17326,6 +17447,34 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ac0b09e7857660247b1477e93731de29, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1 &1437720051 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 1951375231002363255, guid: 196469da6d87b9043b7a7261561efb1e,
+    type: 3}
+  m_PrefabInstance: {fileID: 1963587410}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1437720056
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1437720051}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d62097beafe9634d9fac742ac005b31, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Name: 
+  MaxHealth: 0
+  MoveSpeed: 0
+  Damage: 0
+  health: 0
+  rb: {fileID: 0}
+  XPValue: 0
+  DeathParticles: {fileID: 3489350140759512525, guid: e200f945d72584aa5be5396faed451a4,
+    type: 3}
+  thisSpriteRenderer: {fileID: 0}
 --- !u!1001 &1489555367
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -19310,6 +19459,34 @@ Transform:
   m_Father: {fileID: 172185338}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &1818509049 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 1951375231002363255, guid: 196469da6d87b9043b7a7261561efb1e,
+    type: 3}
+  m_PrefabInstance: {fileID: 323460667}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1818509054
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1818509049}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d62097beafe9634d9fac742ac005b31, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Name: 
+  MaxHealth: 999
+  MoveSpeed: 0
+  Damage: 0
+  health: 0
+  rb: {fileID: 0}
+  XPValue: 0
+  DeathParticles: {fileID: 3489350140759512525, guid: e200f945d72584aa5be5396faed451a4,
+    type: 3}
+  thisSpriteRenderer: {fileID: 0}
 --- !u!1 &1896335502
 GameObject:
   m_ObjectHideFlags: 0
@@ -20140,6 +20317,16 @@ PrefabInstance:
         type: 3}
       propertyPath: m_Name
       value: Focus Door (1)
+      objectReference: {fileID: 0}
+    - target: {fileID: 1951375231002363255, guid: 196469da6d87b9043b7a7261561efb1e,
+        type: 3}
+      propertyPath: m_TagString
+      value: Wall
+      objectReference: {fileID: 0}
+    - target: {fileID: 7319345577122859348, guid: 196469da6d87b9043b7a7261561efb1e,
+        type: 3}
+      propertyPath: MaxHealth
+      value: 999
       objectReference: {fileID: 0}
     - target: {fileID: 8734296311972347452, guid: 196469da6d87b9043b7a7261561efb1e,
         type: 3}
