+++ Assets/Scripts/BlockerDoorScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b3dbc667b5177a45b685a86862fa960
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
